pyproject.toml
[build-system]
requires = ["setuptools>=77", "setuptools-scm[toml]>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "codex-a2a"
dynamic = ["version"]
description = "Codex A2A runtime adapter"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [
    { name = "liujuanjuan1984@Intelligent-Internet" },
]
keywords = ["a2a", "codex", "fastapi", "json-rpc", "sse"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "a2a-sdk==1.0.2",
    "aiosqlite>=0.20,<1.0",
    "fastapi>=0.121,<0.137",
    "httpx>=0.28.1,<0.29",
    "protobuf>=6.33.5,<7.0",
    "pyasn1>=0.6.3,<1.0",
    "python-dotenv>=1.2.2,<2.0",
    "pydantic>=2.6,<3.0",
    "pydantic-settings>=2.2,<3.0",
    "requests>=2.33.0,<3.0",
    "sqlalchemy>=2.0,<3.0",
    "sse-starlette>=2.1,<4.0",
    "uvicorn>=0.29,<0.46",
]

[project.optional-dependencies]
dev = [
    "pip-audit>=2.9,<3.0",
    "pre-commit>=3.7",
    "pygments>=2.20.0,<3.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0,<8.0",
    "ruff>=0.5",
    "mypy>=1.11,<1.21",
]

[project.scripts]
codex-a2a = "codex_a2a.cli:main"

[project.urls]
Homepage = "https://github.com/liujuanjuan1984/codex-a2a"
Repository = "https://github.com/liujuanjuan1984/codex-a2a"
Issues = "https://github.com/liujuanjuan1984/codex-a2a/issues"
Documentation = "https://github.com/liujuanjuan1984/codex-a2a/tree/main/docs"
Releases = "https://github.com/liujuanjuan1984/codex-a2a/releases"
Security = "https://github.com/liujuanjuan1984/codex-a2a/security/policy"

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.uv]
package = true

[tool.setuptools.package-data]
codex_a2a = ["py.typed"]

[tool.setuptools_scm]
fallback_version = "0.0.0+unknown"
tag_regex = "^v?(?P<version>.+)$"

[tool.pytest.ini_options]
addopts = "--cov=src/codex_a2a --cov-branch --cov-report=term-missing:skip-covered"
filterwarnings = [
    "ignore:'HTTP_413_REQUEST_ENTITY_TOO_LARGE' is deprecated. Use 'HTTP_413_CONTENT_TOO_LARGE' instead.:DeprecationWarning:a2a\\.server\\.apps\\.jsonrpc\\.fastapi_app",
]

[tool.coverage.report]
fail_under = 78
show_missing = true
skip_covered = true
